cntd

+++ database/migrations/2019_07_24_165927_create_invoice_requests_table.php
@@ -17,10 +17,13 @@ class CreateInvoiceRequestsTable extends Migration
             $table->increments('id');
 
             $table->unsignedInteger('order_id')->unique();
-            $table->foreign('order_id')
-                ->references('id')
-                ->on('orders')
-                ->onDelete('cascade');
+
+            if (config('database.default') !== 'sqlite') {
+                $table->foreign('order_id')
+                    ->references('id')
+                    ->on('orders')
+                    ->onDelete('cascade');
+            }
 
             $table->dateTime('refused_at')->nullable();
